Implement clean policy factory methods
