+++ src/WiredBrain.Billing/Policies/PolicyExtensions.cs
@@ -1,0 +1,28 @@
+using Microsoft.Extensions.Options;
+using Polly;
+using WiredBrain.Billing.Models;
+
+namespace WiredBrain.Billing.Policies;
+
+public static class PolicyExtensions
+{
+    public static IHttpClientBuilder AddPolicies(this IHttpClientBuilder builder, params Func<ILogger, ResilienceConfig, IAsyncPolicy<HttpResponseMessage>>[] policyFunctions)
+    {
+        if (policyFunctions.Length == 0)
+        {
+            return builder;
+        }
+        return builder.AddPolicyHandler((serviceProvider, _) =>
+            {
+                var logger = serviceProvider.GetRequiredService<ILogger<Program>>();
+                var resilienceConfig = serviceProvider.GetRequiredService<IOptions<ResilienceConfig>>().Value;
+                var policies = policyFunctions.Select(func => func(logger, resilienceConfig)).ToArray();
+                if (policies.Length == 1)
+                {
+                    return policies[0];
+                }
+                var wrapper = Policy.WrapAsync(policies);
+                return wrapper;
+            });
+    }
+}
