+++ src/WiredBrain.Billing/appsettings.Development.json
@@ -11,11 +11,12 @@
   "PaymentService": {
     "BaseUrl": "http://simulated-payments:80/",
     "Resilience": {
-      "TimeoutSeconds": 3,
-      "MaxRetryAttempts": 5,
-      "InitialBackoffSeconds": 0.5,
-      "BackoffMultiplier": 1.5,
-      "JitterFactor": 0.1
+      "TimeoutSeconds": 5,
+      "MaxRetryAttempts": 3,
+      "InitialBackoffSeconds": 1,
+      "ExceptionsAllowedBeforeBreaking": 2,
+      "DurationOfBreakSeconds": 60,
+      "SamplingDurationSeconds": 30
     }
   }
 }
