+++ src/WiredBrain.Billing/appsettings.json
@@ -15,9 +15,9 @@
       "TimeoutSeconds": 5,
       "MaxRetryAttempts": 3,
       "InitialBackoffSeconds": 1,
-      "ExceptionsAllowedBeforeBreaking": 3,
-      "DurationOfBreakSeconds": 30,
-      "SamplingDurationSeconds": 60
+      "ExceptionsAllowedBeforeBreaking": 2,
+      "DurationOfBreakSeconds": 60,
+      "SamplingDurationSeconds": 30
     }
   }
 }
